kaggle-5-day-ai-agents/requirements.txt
google-adk
google-genai
python-dotenv
aiohttp